macos: extract inline title editing to standalone file
